Split color classes in separate files
